app/requirements.txt
azure-functions
scikit-learn
joblib
